refactor: move a pasta utils utilizada pelos testes e tira timeout padrão de alguns testes
